new contracts

+++ Solidity/36_Library.sol
@@ -2,19 +2,19 @@
 pragma solidity >=0.5.0;
 
 library Math {
-    function max(uint x, uint y) internal pure returns (uint) {
+    function max(uint256 x, uint256 y) internal pure returns (uint256) {
         return x >= y ? x : y;
     }
 }
 
 contract Test {
-    function testMax(uint x, uint y) external pure returns (uint) {
+    function testMax(uint256 x, uint256 y) external pure returns (uint256) {
         return Math.max(x, y);
     }
 }
 
-
 library ArrayLib {
+    // prettier-ignore
     function find(uint[] storage arr, uint x) internal view returns (uint) {
         for (uint256 i = 0; i < arr.length; i++) {
             if (arr[i] == x) return i;
@@ -25,11 +25,11 @@ library ArrayLib {
 }
 
 contract TestArray {
-    using ArrayLib for uint[];
+    using ArrayLib for uint256[];
 
-    uint[] public arr = [1, 2, 3];
+    uint256[] public arr = [1, 2, 3];
 
-    function testFind() external view returns (uint i) {
+    function testFind() external view returns (uint256 i) {
         // i = ArrayLib.find(arr, 2);
         i = arr.find(2);
     }
