+++ Solidity/40_DeleteContracts.sol
@@ -7,22 +7,20 @@ pragma solidity >=0.5.0;
     force send ether to any address
 */
 contract DeleteContract {
-    constructor () payable {
-        
-    }
-    
+    constructor() payable {}
+
     function kill() external {
         selfdestruct(payable(msg.sender));
     }
 
-    function testCall() external pure returns (uint) {
+    function testCall() external pure returns (uint256) {
         return 123;
     }
 }
 
 contract Helper {
-    function getBalance() external view returns (uint) {
-        return address(this).balance; 
+    function getBalance() external view returns (uint256) {
+        return address(this).balance;
     }
 
     function kill(DeleteContract _kill) external {
