+++ Solidity/41_PiggyBank.sol
@@ -4,13 +4,12 @@ pragma solidity >=0.5.0;
 contract PiggyBank {
     address public owner = msg.sender;
 
-    event Deposit(uint amount);
-    event Withdraw(uint amount);
+    event Deposit(uint256 amount);
+    event Withdraw(uint256 amount);
 
     receive() external payable {
         emit Deposit(msg.value);
     }
-
 
     function withdraw() external {
         require(msg.sender == owner, "not owner");
